🧹 Refactor Monolithic Server Initialization: extract auth middleware
